refactor: new block emitter into separate files
